+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPPVWAGA5BH9BHVAWXMZT5S9"
+        "01KPQH7SQQT9C2B1WYXYV8P3TV"
     )
     r"""The unique identifier of the data source"""