+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPPVWAG7TDJZVZA5CR2PZ5WT"
+        "01KPQH7SQNFVV4GJ7XQPSJTD4F"
     )
     r"""The unique identifier of the data source"""