+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPPVWAG90PTJPA76VHKJBS7H"
+        "01KPQH7SQPN3NT2YMAAN93PF6S"
     )
     r"""The unique identifier of the data source"""