+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -154,7 +154,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-04-21T01:53:17.821Z")
+    updated: Optional[datetime] = parse_datetime("2026-04-21T08:06:34.615Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")