+++ packages/orq-rc/RELEASES.md
@@ -7119,3 +7119,13 @@ Based on:
 - [python v4.8.0-rc.34] packages/orq-rc
 ### Releases
 - [PyPI v4.8.0-rc.34] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.34 - packages/orq-rc
+
+## 2026-04-21 08:09:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.8.0-rc.35] packages/orq-rc
+### Releases
+- [PyPI v4.8.0-rc.35] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.35 - packages/orq-rc